+++ packages/atomicdex-app/app/containers/DexPage/components/AmountSection.js
@@ -29,10 +29,12 @@ import validate from '../../../components/Form/validate';
 import type { BuyCoinPayload } from '../schema';
 import { calculateDexfee } from '../utils';
 import {
-  AUTO_HIDE_SNACKBAR_TIME,
   STATE_SWAPS,
   NA,
   FINISHED_SWAPS_STATE
+} from '../../../constants';
+import {
+  AUTO_HIDE_SNACKBAR_TIME,
 } from '../constants';
 import {
   loadBuyCoin,
