+++ packages/atomicdex-app/app/containers/DexPage/reducer.js
@@ -4,6 +4,17 @@ import { handleActions } from 'redux-actions';
 import { LOGOUT } from '../App/constants';
 // import logger from '../../logger';
 import {
+  // NEGOTIATED_SWAPS_STATE,
+  STARTED_SWAPS_STATE,
+  TAKER_PAYMENT_SENT_SWAPS_STATE,
+  TAKER_FEE_SENT_SWAPS_STATE,
+  MAKER_PAYMENT_RECEIVED_SWAPS_STATE,
+  MAKER_PAYMENT_SPENT_SWAPS_STATE,
+  TAKER_PAYMENT_SPENT_SWAPS_STATE,
+  FINISHED_SWAPS_STATE
+} from '../../constants';
+
+import {
   LOAD_PRICES,
   LOAD_BEST_PRICE,
   LOAD_PRICES_SUCCESS,
@@ -25,15 +36,7 @@ import {
   SELECT_COIN_MODAL_CLICK,
   COIN_PAYMENT_SELECT,
   JOYRIDE_OPEN,
-  JOYRIDE_CLOSE,
-  STARTED_SWAPS_STATE,
-  // NEGOTIATED_SWAPS_STATE,
-  TAKER_FEE_SENT_SWAPS_STATE,
-  MAKER_PAYMENT_RECEIVED_SWAPS_STATE,
-  TAKER_PAYMENT_SENT_SWAPS_STATE,
-  MAKER_PAYMENT_SPENT_SWAPS_STATE,
-  TAKER_PAYMENT_SPENT_SWAPS_STATE,
-  FINISHED_SWAPS_STATE
+  JOYRIDE_CLOSE
 } from './constants';
 
 // The initial state of the App
